v5

+++ hw1/main.cpp
@@ -5,18 +5,18 @@
 
 int main(){
     std::ifstream input_file("in.txt");
-    float xn;
-    float yn;
+    double xn;
+    double yn;
     input_file >> xn >> yn;
-    float xr=xn;
-    float xl=xn;
-    float yr=yn;
-    float yl=yn;
+    double xr=xn;
+    double xl=xn;
+    double yr=yn;
+    double yl=yn;
     double a = atan(xn/yn);
     while (input_file){
         input_file >> xn >> yn;
-        float x1=xn*cos(a)+yn*sin(a);
-        float y1=yn*cos(a)-xn*sin(a);
+        double x1=xn*cos(a)+yn*sin(a);
+        double y1=yn*cos(a)-xn*sin(a);
         if (x1>xr){
             x1=xr;
             y1=yr;
